Working example of project saving/loading
